Extend to Euler system. Clean ups.
