+++ pyfr/solvers/euler/inters.py
@@ -1,12 +1,8 @@
-import numpy as np
-
-from pyfr.mpiutil import mpi, scal_coll
-from pyfr.quadrules.surface import SurfaceIntegrator
 from pyfr.solvers.baseadvec import (BaseAdvectionIntInters,
                                     BaseAdvectionMPIInters,
                                     BaseAdvectionBCInters)
-from pyfr.util import first
-from pyfr.writers.csv import CSVStream
+from pyfr.solvers.euler.mixins import (MassFlowBCMixin, NSCBCMixin,
+                                       PressureBCMixin)
 
 
 class TplargsMixin:
@@ -105,190 +101,41 @@ class EulerSlpAdiaWallBCInters(EulerBaseBCInters):
     type = 'slp-adia-wall'
 
 
-class ControlledBCMixin:
-    def __init__(self, be, lhs, elemap, cfgsect, cfg, bccomm):
-        super().__init__(be, lhs, elemap, cfgsect, cfg, bccomm)
-
-        self.c |= self._exp_opts(
-            ['rho', 'u', 'v', 'w'][:self.ndims + 1], lhs
-        )
-
-        self.con_to_pri = first(elemap.values()).con_to_pri
-
-        self.tstart = cfg.getfloat(cfgsect, 'tstart', 0.0)
-        self.nsteps = cfg.getint(cfgsect, 'nsteps', 100)
-
-        opts = self._eval_opts(self._target_opts)
-        self.target, self.alpha, self.eta = opts
-
-        self.set_external('ic', 'scalar fpdtype_t')
-        self.set_external('im', 'scalar fpdtype_t')
-
-        self.surf_int = SurfaceIntegrator(cfg, cfgsect, elemap, lhs)
-
-        self._init_extra(cfg, cfgsect)
-
-        if cfg.hasopt(cfgsect, 'file') and bccomm.rank == 0:
-            fname = cfg.get(cfgsect, 'file')
-            nflush = cfg.getint(cfgsect, 'flushsteps', 10)
-            self.csv = CSVStream(fname, header=self._csv_header, nflush=nflush)
-        else:
-            self.csv = None
-
-    def _init_extra(self, cfg, cfgsect):
-        pass
-
-    def _interp_face(self, solns):
-        for etype, fidx in self.surf_int.m0:
-            m0 = self.surf_int.m0[etype, fidx]
-            nfpts, nupts = m0.shape
-
-            eidxs = self.surf_int.eidxs[etype, fidx]
-            uupts = solns[etype][:, :, eidxs]
-
-            ufpts = m0 @ uupts.reshape(nupts, -1)
-            ufpts = ufpts.reshape(nfpts, self.nvars, -1).swapaxes(0, 1)
-
-            qwts = self.surf_int.qwts[etype, fidx]
-            norms = self.surf_int.norms[etype, fidx]
-
-            yield ufpts, qwts, norms
-
-    def setup(self, sdata, prevcfg):
-        sect_eq = (prevcfg is not None and
-                   self.cfg.sect_eq(prevcfg, self.cfgsect))
-
-        if sdata is not None and sdata[4] != 0 and sect_eq:
-            self.interp_c = sdata[0]
-            self.interp_m = sdata[1]
-            self.meas_avg = sdata[2]
-            self.tprev = sdata[3]
-            self.nstep_counter = sdata[4]
-        else:
-            self.interp_c = self._default_interp_c()
-            self.interp_m = 0.0
-            self.meas_avg = 0.0
-            self.tprev = None
-            self.nstep_counter = 0
-
-    @classmethod
-    def preparefn(cls, bciface, mesh, elemap):
-        if bciface:
-            return bciface.prepare
-        else:
-            return None
-
-    def prepare(self, system, ubank, t, kerns):
-        update = self.nstep_counter % self.nsteps == 0
-        if (update or not self.tprev) and t >= self.tstart:
-            solns = dict(zip(system.ele_types, system.ele_scal_upts(ubank)))
-            meas = self._measure(solns)
-
-            if not self.tprev:
-                self.meas_avg = meas
-                self.tprev = t
-            else:
-                a = self.alpha
-                self.meas_avg = a*meas + (1 - a)*self.meas_avg
-                dt = t - self.tprev
-                self.tprev = t
-
-                # Current Riemann invariant pressure
-                p0 = self.interp_m*t + self.interp_c
-
-                # Compute corrected Riemann pressure
-                p1 = self._correction(p0, dt)
-
-                # Update interpolation coefficients
-                self.interp_m = (p1 - p0) / dt
-                self.interp_c = p0 - self.interp_m*t
-
-                # Log to CSV
-                if self.csv:
-                    self.csv(t, self.meas_avg, p1)
-
-        # Bind interpolation to kernels
-        for k in kerns.values():
-            k.bind(ic=self.interp_c, im=self.interp_m)
-
-        self.nstep_counter += 1
-
-    @classmethod
-    def serialisefn(cls, bciface, prefix, srl):
-        sfn = lambda: np.void(
-            (bciface.interp_c, bciface.interp_m, bciface.meas_avg,
-             bciface.tprev or 0, bciface.nstep_counter),
-            dtype='f8,f8,f8,f8,i8'
-        )
-        srl.register(prefix, sfn if bciface else None)
-
-
-class MassFlowBCMixin(ControlledBCMixin):
-    _target_opts = ['mass-flow-rate', 'alpha', 'eta']
-    _csv_header = 't,mf,pbc'
-
-    def _init_extra(self, cfg, cfgsect):
-        self._qwts_norms = []
-
-        for etype, fidx in self.surf_int.m0:
-            qwts = self.surf_int.qwts[etype, fidx]
-            norms = self.surf_int.norms[etype, fidx]
-            self._qwts_norms.append(norms*qwts[:, None])
-
-    def _default_interp_c(self):
-        return self._eval_opts(['p'])[0]
-
-    def _measure(self, solns):
-        mf = 0.0
-
-        for (u, *_), qn in zip(self._interp_face(solns), self._qwts_norms):
-            mf += np.einsum('hij,hij', u[1:-1], qn)
-
-        return scal_coll(self.bccomm.Allreduce, mf, op=mpi.SUM)
-
-    def _correction(self, p0, dt):
-        return p0 + dt*self.eta*(1 - self.target / self.meas_avg)
-
-
 class EulerCharRiemInvMassFlowBCInters(MassFlowBCMixin, EulerBaseBCInters):
     type = 'char-riem-inv-mass-flow'
 
 
-class PressureBCMixin(ControlledBCMixin):
-    _target_opts = ['pressure', 'alpha', 'eta']
-    _csv_header = 't,p,pbc'
-
-    def _init_extra(self, cfg, cfgsect):
-        area = 0.0
-        self._qwts_nmag = []
-
-        all_qwts = self.surf_int.qwts.values()
-        all_norms = self.surf_int.norms.values()
-        for qwts, norms in zip(all_qwts, all_norms):
-            nmag = np.sqrt(np.einsum('hij,hij->ij', norms, norms))
-            qwts_nmag = nmag*qwts[:, None]
-            self._qwts_nmag.append(qwts_nmag)
-            area += qwts_nmag.sum()
-
-        self.area = scal_coll(self.bccomm.Allreduce, area, op=mpi.SUM)
-
-    def _default_interp_c(self):
-        return self.target
-
-    def _measure(self, solns):
-        p_num = 0.0
-
-        for (u, *_), qnmag in zip(self._interp_face(solns), self._qwts_nmag):
-            p = self.con_to_pri(u, self.cfg)[-1]
-            p_num += np.einsum('ij,ij', p, qnmag)
-
-        p_num = scal_coll(self.bccomm.Allreduce, p_num, op=mpi.SUM)
-
-        return p_num / self.area
-
-    def _correction(self, p0, dt):
-        return p0 + dt*self.eta*(self.target - self.meas_avg)
-
-
 class EulerCharRiemInvPressureBCInters(PressureBCMixin, EulerBaseBCInters):
     type = 'char-riem-inv-pressure'
+
+
+class EulerNSCBCSubOutFpBCInters(NSCBCMixin, EulerBaseBCInters):
+    type = 'sub-out-nscbc-fp'
+    waves = ['acoustic-']
+
+    def __init__(self, be, lhs, elemap, cfgsect, cfg, bccomm):
+        super().__init__(be, lhs, elemap, cfgsect, cfg, bccomm)
+
+        for face in self.nscbc_faces:
+            self.c |= self._exp_opts_ele(['p'], face.lhs, face.extern_args,
+                                         face.extern_vals)
+        self.c['K_p'] = self.cfg.getfloat(cfgsect, 'K_p', default=1.0)
+
+
+class EulerNSCBCSubInNRIBCInters(NSCBCMixin, EulerBaseBCInters):
+    type = 'sub-in-nscbc-nri'
+    flip_norm = True
+    waves = ['entropy', 'vortical', 'acoustic+']
+
+    def __init__(self, be, lhs, elemap, cfgsect, cfg, bccomm):
+        super().__init__(be, lhs, elemap, cfgsect, cfg, bccomm)
+
+        force = ['u_a', 'du_a_dt', 'u_v', 'du_v_dt']
+        for face in self.nscbc_faces:
+            self.c |= self._exp_opts_ele(['rho', 'un'], face.lhs,
+                                         face.extern_args, face.extern_vals)
+            self.c |= self._exp_opts_ele(force, face.lhs, face.extern_args,
+                                         face.extern_vals,
+                                         default={f: 0.0 for f in force})
+        for n in ['isen', 'ut']:
+            self.c[f'K_{n}'] = self.cfg.getfloat(cfgsect, f'K_{n}', default=1.0)
