+++ pyfr/solvers/navstokes/inters.py
@@ -1,12 +1,10 @@
-from collections import defaultdict
-
 import numpy as np
 
-from pyfr.readers.native import Connectivity
 from pyfr.solvers.baseadvecdiff import (BaseAdvectionDiffusionBCInters,
                                         BaseAdvectionDiffusionIntInters,
                                         BaseAdvectionDiffusionMPIInters)
-from pyfr.solvers.euler.inters import MassFlowBCMixin, PressureBCMixin
+from pyfr.solvers.euler.mixins import (MassFlowBCMixin, NSCBCMixin,
+                                       PressureBCMixin)
 
 
 class TplargsMixin:
@@ -221,171 +219,35 @@ class NavierStokesCharRiemInvPressureBCInters(PressureBCMixin,
     cflux_state = 'ghost'
 
 
-class NSCBCMixin:
+class NavierStokesNSCBCSubOutFpBCInters(NSCBCMixin, NavierStokesBaseBCInters):
+    type = 'sub-out-nscbc-fp'
+    waves = ['acoustic-']
     _nscbc_kern = 'pyfr.solvers.navstokes.kernels.bccflux_nscbc'
 
     def __init__(self, be, lhs, elemap, cfgsect, cfg, bccomm):
         super().__init__(be, lhs, elemap, cfgsect, cfg, bccomm)
 
-        # NSCBC replaces the standard comm_flux with nscbc_flux
-        self.kernels.pop('comm_flux', None)
-        self._be.pointwise.register(self._nscbc_kern)
-
-        self._tplargs_efp = defaultdict(dict)
-        self._external_args_efp = defaultdict(dict)
-        self._external_vals_efp = defaultdict(dict)
-        self._scal_upts = defaultdict(dict)
-        self._scal_fpts = defaultdict(dict)
-        self._grad_upts = defaultdict(dict)
-        self._smats_upts = defaultdict(dict)
-        self._jacs_facefpts = defaultdict(dict)
-        self._dim_lhs = defaultdict(dict)
-
-        # Register NSCBC kernel under a different name so it can be scheduled after other BCs
-        self.kernels['nscbc_flux'] = lambda: self.gen_nscbc_kerns()
-
-        # Create sub-Connectivity for each element-face pair
-        self.ef_pairs = []
-        self._lhs_efp = defaultdict(dict)
-        for etype, fidx, eidxs in lhs.items():
-            self.ef_pairs.append((etype, fidx))
-            cidx = next(c for c, (et, fi) in lhs.cidxmap.items()
-                        if et == etype and fi == fidx)
-            mask = lhs.cidxs == cidx
-            self._lhs_efp[etype][fidx] = Connectivity(
-                lhs.cidxs[mask], lhs.eidxs[mask], lhs.cidxmap
-            )
-
-        for etype, fidx in self.ef_pairs:
-            lhs_efp = self._lhs_efp[etype][fidx]
-            self._dim_lhs[etype][fidx] = len(lhs_efp)
-
-            self._external_args_efp[etype][fidx] = self._external_args.copy()
-            self._external_vals_efp[etype][fidx] = self._external_vals.copy()
-
-            ele = self.elemap[etype]
-            basis = ele.basis
-            nupts = basis.nupts
-            nfpts = basis.nfpts
-            nfacefpts = basis.nfacefpts[fidx]
-            ndims = self.ndims
-            facefpts = basis.facefpts[fidx]
-            fptidx = [i for j in basis.facefpts for i in j]
-            intfpts = [i for i in fptidx if i not in facefpts]
-
-            # Reference normal directions for computing physical normals
-            magnl = np.linalg.norm(basis.norm_fpts, axis=-1)
-            norms = basis.norm_fpts[facefpts] / magnl[facefpts, None]
-            if '-in-' in self.type:
-                norms = -norms
-
-            # Correction function matrices for FR inversion
-            # m11[i,j] is the divergence of correction function j at flux point i
-            GB = basis.m11[np.ix_(facefpts, facefpts)]
-            GI = basis.m11[np.ix_(facefpts, intfpts)]
-            GB_inv = np.linalg.inv(GB)
-
-            self._tplargs_efp[etype][fidx] = self._tplargs | dict(
-                nupts=nupts, nfpts=nfpts, nfacefpts=nfacefpts,
-                nintfacefpts=nfpts - nfacefpts, facefpts=facefpts,
-                fptidx=fptidx, intfpts=intfpts,
-                m0=basis.m0[facefpts],
-                m2=basis.m2.reshape(nfpts, ndims, nupts),
-                m12=basis.m12[facefpts], norm_ref=norms,
-                GB_inv=GB_inv, GB_inv_GI=GB_inv @ GI,
-                decomp_type=self.decomp_type,
-            )
-
-            self._scal_upts[etype][fidx] = self._scal_upts_view(
-                lhs_efp, '_get_scal_upts_cpy_ewise')
-            self._scal_fpts[etype][fidx] = self._scal_fpts_view(
-                lhs_efp, '_get_scal_fpts_ewise')
-            self._grad_upts[etype][fidx] = self._grad_upts_view(
-                lhs_efp, '_get_grad_upts_ewise')
-
-            # Collect smats at upts: (ndims, nupts, ndims, neles) -> (nupts, ndims*ndims, neles_efp)
-            _, _, eidxs_efp = next(lhs_efp.items())
-            smats = ele.smat_at_np('upts')[:, :, :, eidxs_efp]
-            smats = smats.transpose(1, 0, 2, 3).reshape(nupts, ndims*ndims, -1)
-            self._smats_upts[etype][fidx] = self._be.const_matrix(smats)
-
-            # Collect jacs at face fpts: (nfacefpts, neles_efp)
-            jacs = 1.0 / ele.rcpdjac_at_np('fpts')[facefpts][:, eidxs_efp]
-            self._jacs_facefpts[etype][fidx] = self._be.const_matrix(jacs)
-
-    def gen_nscbc_kerns(self):
-        kerns = []
-        for etype, fidx in self.ef_pairs:
-            kerns.append(self._be.kernel(
-                'bccflux_nscbc',
-                tplargs=self._tplargs_efp[etype][fidx],
-                dims=[self._dim_lhs[etype][fidx]],
-                extrns=self._external_args_efp[etype][fidx],
-                u_upts=self._scal_upts[etype][fidx],
-                u_fpts=self._scal_fpts[etype][fidx],
-                gradu_upts=self._grad_upts[etype][fidx],
-                smats_upts=self._smats_upts[etype][fidx],
-                jacs_ffpts=self._jacs_facefpts[etype][fidx],
-                **self._external_vals_efp[etype][fidx]))
-
-        return self._be.unordered_meta_kernel(kerns)
+        for face in self.nscbc_faces:
+            self.c |= self._exp_opts_ele(['p'], face.lhs, face.extern_args,
+                                         face.extern_vals)
+        self.c['K_p'] = self.cfg.getfloat(cfgsect, 'K_p', default=1.0)
 
 
-class NSCBCSubOutFpBCInters(NSCBCMixin, NavierStokesBaseBCInters):
-
-    type = 'sub-out-nscbc-fp'
-    decomp_type = 'normal'
-
-    def __init__(self, be, lhs, elemap, cfgsect, cfg, bccomm):
-        super().__init__(be, lhs, elemap, cfgsect, cfg, bccomm)
-
-        for etype, fidx in self.ef_pairs:
-            lhs_efp = self._lhs_efp[etype][fidx]
-            self.c |= self._exp_opts_ele(['p'], lhs_efp,
-                                         self._external_args_efp[etype][fidx],
-                                         self._external_vals_efp[etype][fidx])
-        self.c['K_p'] = self.cfg.getfloat(cfgsect, 'K_p', default=0.25)
-
-class NSCBCSubInFrvBCInters(NSCBCMixin, NavierStokesBaseBCInters):
-
-    type = 'sub-in-nscbc-frv'
-    decomp_type = 'cartesian'
-
-    def __init__(self, be, lhs, elemap, cfgsect, cfg, bccomm):
-        super().__init__(be, lhs, elemap, cfgsect, cfg, bccomm)
-        for etype, fidx in self.ef_pairs:
-            lhs_efp = self._lhs_efp[etype][fidx]
-
-            self.c |= self._exp_opts_ele(
-                ['rho', 'u', 'v', 'w'][:self.ndims + 1], lhs_efp,
-                self._external_args_efp[etype][fidx],
-                self._external_vals_efp[etype][fidx],
-            )
-        for i in ['rho', 'u', 'v', 'w'][:self.ndims + 1]:
-            self.c[f'K_{i}'] = self.cfg.getfloat(cfgsect, f'K_{i}', default=0.25)
-
-class NSCBCSubInNRIBCInters(NSCBCMixin, NavierStokesBaseBCInters):
-
+class NavierStokesNSCBCSubInNRIBCInters(NSCBCMixin, NavierStokesBaseBCInters):
     type = 'sub-in-nscbc-nri'
-    decomp_type = 'normal'
+    flip_norm = True
+    waves = ['entropy', 'vortical', 'acoustic+']
+    _nscbc_kern = 'pyfr.solvers.navstokes.kernels.bccflux_nscbc'
 
     def __init__(self, be, lhs, elemap, cfgsect, cfg, bccomm):
         super().__init__(be, lhs, elemap, cfgsect, cfg, bccomm)
 
         force = ['u_a', 'du_a_dt', 'u_v', 'du_v_dt']
-        for etype, fidx in self.ef_pairs:
-            lhs_efp = self._lhs_efp[etype][fidx]
-
-            self.c |= self._exp_opts_ele(
-                ['rho', 'un'], lhs_efp,
-                self._external_args_efp[etype][fidx],
-                self._external_vals_efp[etype][fidx],
-            )
-            self.c |= self._exp_opts_ele(
-                force, lhs_efp,
-                self._external_args_efp[etype][fidx],
-                self._external_vals_efp[etype][fidx],
-                default={f: 0.0 for f in force},
-            )
-        for i in ['ac', 'ut']:
-            self.c[f'K_{i}'] = self.cfg.getfloat(cfgsect, f'K_{i}', default=0.25)
+        for face in self.nscbc_faces:
+            self.c |= self._exp_opts_ele(['rho', 'un'], face.lhs,
+                                         face.extern_args, face.extern_vals)
+            self.c |= self._exp_opts_ele(force, face.lhs, face.extern_args,
+                                         face.extern_vals,
+                                         default={f: 0.0 for f in force})
+        for n in ['isen', 'ut']:
+            self.c[f'K_{n}'] = self.cfg.getfloat(cfgsect, f'K_{n}', default=1.0)
